09/19/16
* FBMovement script: changed "Update" to "FixedUpdate" since it only
deals with chaning the FB's velocity
* Jump script: changed "Update" to "FixedUpdate" since it only deals
with changing Surfer Girl's velocity
* TikiAnimationManager script: deleted code which was not being used.
* ShootFB script: Deleted the entire "Update" function and replaced it
with a GameManagerEvent to determine when to shoot the fireball. This is
in an attemp to improve code efficiency. I do not see a point in
checking to see if a fireball should be shot every frame update.
* GameManager script: Added delegate and event functions to broadcast
when the GameState has changed. This is in an attempt to improve code
efficiency so that each game objects will not have to constantly check
the game state on each frame. The GameState function will broadcast
this. Deleted variables and functions which were replaced by these
changes.
* FBAnimationManager script: replaced the "Update" function with
GameManager Events for previously stated reasons.
* CollisionState script: Deleted the variable "_blockedFB" because it
was no longer needed with all the previous changes.
* PlayerAnimationManager script: Deleted the "Update" function due to
previous mentioned changes.
* Changed almost all scripts similarly to the previously mentioned for
the same reasons.